scripts/streaming_consumer/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
azure-core==1.28.0
azure-eventhub==5.11.3
azure-eventhub-checkpointstoreblob-aio==1.1.4
azure-identity==1.13.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.2
frozenlist==1.4.0
idna==3.4
isodate==0.6.1
msal==1.23.0
msal-extensions==1.0.0
msrest==0.7.1
multidict==6.0.4
oauthlib==3.2.2
portalocker==2.7.0
pycparser==2.21
PyJWT==2.8.0
python-dotenv==1.0.0
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
typing_extensions==4.7.1
urllib3==2.0.4
yarl==1.9.2
